legal-doc-simplifier/requirements.txt
Flask                         
huggingface_hub
python-dotenv
torch
transformers
python-docx
pymupdf  
pytesseract
Pillow
werkzeug
requests
IndicTransToolkit
json
threading
sentencepiece